MHWD refactoring, part 1 (#6)
* mhwd: Handle pacman errors in function

* mhwd: Drop 32 bit support

We don't intend to support 32 bit systems

* mhwd: Use pacman search for kernels instead of grep

* mhwd: Refactor of post/pre functions checks

* mhwd: Use pacman search for REMOVEPKGS instead of grep

* mhwd: Make multilib autodetection

* mhwd: Use PMCONFIG to check multilib

* mhwd: Replace $INSTALL and $REMOVE by $ACTION

* mhwd: Replace all checks on empty by the built-in option in test

* mhwd: Hotfix

Fixes: https://github.com/CachyOS/mhwd-cachyos/pull/6/commits/cad1796e5bbe02a1f664a639338d8d6481118cff

* mhwd: Get rid of CONKMOD

Pacman itself can solve conflicts, there is no point in specifying a
possible list of package conflicts in the profiles, which list can be
quite large. For profile conflicts we have MHWDCONFLICT.

This commit requires changes in mhwd-db.

* mhwd: Return --noconfirm instead of pipe

* mhwd: Fix exit code check

* mhwd: Fix exit code check (2)

* mhwd: Don't remove packages in OPTREMDEPENDS

Just Install them as dependencies. In case they are not needed, they
will become orphan packages and can be safely removed by the user.

* mhwd: Hotfix

* mhwd: Fix removal conflicts