PRIME: Remove libva-nvidia-driver from installable packages (#132)
On hybrid systems (only AMD so far), the nvidia-vaapi driver can cause issues with a few software,
namely Spectacle. When this is installed on the system, it will be the first driver picked for encoding.
There are no issues with this on first launch. However subsequent launches will **not** reinitalize
the VAAPI driver, making Spectacle have a "lag spike" when taking region screenshots.

Since the environmental variable to force the use of nvidia-vaapi has been removed, I think this is
a reasonable fix. Users who want nvidia-vaapi can grab the package manually.

Signed-off-by: Eric Naim <dnaim@cachyos.org>