profiles/handheld: Do not mask any jupiter-* services for non Steam Deck
As far as the bios and controller update are concerned, their respective
scripts will check if the device is a Steam Deck or not, and early exit
if its the latter. jupiter-fan-control is not installed in these
devices, so we do not have to mask it either, and a manual enable is
also necessary if for some reason a user decides to install the
package.

Signed-off-by: Eric Naim <dnaim@cachyos.org>