profiles/handheld: Add support for Lenovo Legion Go S
This adds support for the Lenovo Legion Go S. This device comes with either a Phoenix APU or a Rembrandt APU. Hence, the profile name chosen is prefixed with `amd`, similar to what was done for the MSI Claw(s). Support in `inputplumber`, `steamos-manager` and `steamos-powerbuttond` is already available as this is an officialy supported device in SteamOS. The only remaining thing we need to add is the HID driver for this device, which has been added as of c4aab791b4c0e3747a62b6cf686dc5941e907938 and should arrive in 6.15.4.

Signed-off-by: Eric Naim <dnaim@cachyos.org>