Revert "profiles/handheld: Install audio profile for the Lenovo Legion Go" (#181)
This reverts commit 1a6b636500e71b62e36a8818693e893a48c023c5.

A lot of reports came in from users now that the audio profile is
causing severe crackling. Workarounds include setting a higher TDP limit
or bumping the minimum quant to 512/48000, but that didn't mitigate the
issue completely and we should never tell users to use a higher TDP just
so that audio works correctly.

Signed-off-by: Eric Naim <dnaim@cachyos.org>