profiles/graphic_drivers: Drop 390.xx profile for Fermi GPUs
TL;DR for Fermi users: Use Nouveau or buy a new GPU.

We’ve been coming back to this for quite a while now, and I was hoping
it could wait until the end of year, but it looks like the last kernel
update made some changes. In short, Arch kernel no longer provides a
patch that disables simplefb when using ``nvidia-drm.modeset=1`` [1],
because the issue related to it has been fixed in the latest driver, but
it breaks 390.xx [2][3] and naturally no one will fix this for the EOL
driver.

We also can’t give up using ``nvidia-drm.modeset=1``, even though 390.xx
doesn’t need this parameter to work with certain features, but it has
become a strict requirement for multi-monitor configurations from 6.12
patch (current LTS) [4], and author of this patch has explicitly stated
that it is not going to implement this differently than with
``nvidia-drm.modeset=1`` [5].

Using ``initcall_blacklist=simpledrm_platform_driver_init`` kernel
parameter also does not help 390.xx users for an unknown reason. [6]

So, we have no choice but to abandon 390.xx or leave it in its current
broken state for a number of users, which can create a negative
impression of CachyOS.

In addition, 390.xx has a number of issues that cannot be fixed by
simply patching the kernel modules:

- Lack of Wayland support. 390.xx is the one that doesn’t even support
  EGLStreams, which essentially means that any Wayland session will not
  work on it. Given the current trends in transition of desktop
  environments to Wayland, this is also a reason for not supporting
  390.xx, because we do not have a mechanism to notify users about
  installing an unsupported combination of driver and environment, in
  consequence user who was not aware of such details can install
  wlroots-based compositor and get a broken installation. Also remember
  that GNOME 49 is coming out this month, and it will have X11 support
  disabled by default.

- No Vulkan support. In the current situation, this means that you will
  not be able to play games under Linux normally, as most of them work
  with normal performance thanks to DXVK and VKD3D. WineD3D with OpenGL
  has much worse performance, but even this is no longer possible due to
  OpenGL issue with Wine's WoW64 mode, which was recently adopted by
  Arch.

- Lack of support for PRIME Offload makes normal use of laptops with
  hybrid graphics impossible and greatly impairs autonomy. Your dGPU
  will be always up and running, even when you don’t need it. The use of
  Bumblebee is highly discouraged as it is abandoned and has a number of
  drawbacks.

- No video acceleration support via VA-API. nvidia-vaapi-driver requires
  at least 470.xx driver.

- Security. Since the end of support for 390.xx, a number of
  vulnerabilities have been found in NVIDIA drivers, and we don’t
  know which ones affect 390.xx, and even if we did, there’s no way to
  fix them in closed blobs.

Given the above, I am relieving myself of burden of maintaining 390.xx
in CachyOS. For end users it is recommended to switch to using Nouveau,
or install 390.xx drivers yourself from AUR, but in this case we will
not provide any support in case of issues related to 390.xx. Although
Nouveau has slightly worse performance, it is still supported, works
with Wayland and supports PRIME Offload. I will also try to spend time
on the backport of Nouveau fixes.

[1] - https://github.com/archlinux/linux/releases/tag/v6.16.1-arch1
[2] - https://aur.archlinux.org/packages/nvidia-390xx-dkms#comment-1036972
[3] - https://aur.archlinux.org/packages/nvidia-390xx-dkms#comment-1038038
[4] - https://aur.archlinux.org/cgit/aur.git/tree/kernel-6.12.patch?h=nvidia-390xx-utils#n63
[5] - https://gist.github.com/joanbm/a6d3f7f873a60dec0aa4a734c0f1d64e?permalink_comment_id=5254092#gistcomment-5254092
[6] - https://bbs.archlinux.org/viewtopic.php?pid=2247931#p2247931

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>